financial-qa-backend/requirements.txt
# Core dependencies
fastapi>=0.104.0,<0.116.0
uvicorn>=0.23.2,<0.35.0
python-dotenv>=1.0.0
pydantic>=2.4.2,<3.0.0

# LangChain and related packages
langchain>=0.0.335,<0.4.0
langchain-openai>=0.0.2,<0.4.0
langchain-text-splitters>=0.0.1,<0.4.0
langchain-community>=0.0.1,<0.4.0
langchain-core>=0.1.0,<0.4.0

# Vector database
pinecone-client>=2.2.2
pinecone>=6.0.0

# OpenAI
openai>=1.3.0,<2.0.0

# Firebase/Firestore
firebase-admin>=6.2.0
google-cloud-firestore>=2.13.1

# ML and NLP
sentence-transformers>=2.2.2

# Utilities
coloredlogs>=15.0.1
httpx>=0.24.1
requests>=2.31.0
schedule>=1.2.1
beautifulsoup4>=4.12.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.1

# Optional: For development
black>=23.7.0
isort>=5.12.0
mypy>=1.5.1 